pyproject.toml
[tool.poetry]
name = "dashboard-src"
version = "0.1.0"
description = ""
authors = ["Kojo <boakyeni@usc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
django-phonenumber-field = "^7.2.0"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
phonenumbers = "^8.13.26"
django-scheduler = "^0.10.1"
requests = "^2.31.0"
sorl-thumbnail = "^12.10.0"
jsonfield = "^3.1.0"
pillow = "^10.1.0"
premailer = "^3.10.0"
mysqlclient = "^2.2.1"
jsonify = "^0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
